实现了参数管理和标定 实现了多选搜索界面

+++ BLEScaler/src/com/blescaler/worker/ScalerParam.java
@@ -86,7 +86,7 @@ public class ScalerParam {
 	public int getNov() {
 		return nov;
 	}
-	public void setNov(byte nov) {
+	public void setNov(int nov) {
 		this.nov = nov;
 	}
 	public byte getZerotrack() {
